requirements.txt
openai-whisper==20240930
google-generativeai==0.7.2
numpy==2.0.2
playsound==1.3.0
pyttsx3==2.98
scipy==1.14.1
sounddevice==0.5.1
Flask-Cors==4.0.0
requests==2.31.0
pydub==0.25.1